DataAPIService/requirements.txt
fastapi==0.115.8
passlib==1.7.4
pydantic==2.10.6
PyJWT==2.10.1
python-dotenv==1.0.1
redis==5.2.1
Requests==2.32.3
uvicorn==0.34.0
